+++ src/core/cache/never/index.ts
@@ -22,10 +22,19 @@ export * from 'core/cache/interface';
  * Loopback class for a cache data structure
  */
 export default class NeverCache<V = any, K = any> implements Cache<V, K> {
+	/** @see [[Cache.size]] */
+	get size(): number {
+		return this.storage.size;
+	}
+
 	/**
 	 * Cache object
 	 */
 	protected readonly storage: Map<K, V> = new Map();
+
+	[Symbol.iterator](): IterableIterator<K> {
+		return this.keys();
+	}
 
 	/** @see [[Cache.has]] */
 	has(key: K): boolean {
@@ -48,8 +57,18 @@ export default class NeverCache<V = any, K = any> implements Cache<V, K> {
 	}
 
 	/** @see [[Cache.keys]] */
-	keys(): Iterator<K> {
+	keys(): IterableIterator<K> {
 		return this.storage.keys();
+	}
+
+	/** @see [[Cache.values]] */
+	values(): IterableIterator<V> {
+		return this.storage.values();
+	}
+
+	/** @see [[Cache.entries]] */
+	entries(): IterableIterator<[K, V]> {
+		return this.storage.entries();
 	}
 
 	/** @see [[Cache.clear]] */